# openapi/swagger.yaml
components:
  schemas:
    internal_httpapi.AcceptInviteRequest:
      properties:
        password:
          type: string
        token:
          type: string
      type: object
...
    internal_httpapi.AuthTokensResponse:
      properties:
        accessToken:
          type: string
        backupCodes:
          items:
            type: string
          type: array
          uniqueItems: false
        user:
          $ref: '#/components/schemas/sudoStream_internal_auth.PublicUser'
      type: object
    internal_httpapi.ChangeEmailRequest:
      properties:
        newEmail:
          type: string
        password:
          type: string
        totp:
          type: string
      type: object
    internal_httpapi.ChangePasswordRequest:
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
        totp:
          type: string
      type: object
...
    internal_httpapi.DisableTwoFactorRequest:
      properties:
        code:
          type: string
        password:
          type: string
      type: object
    internal_httpapi.ErrorResponse:
      properties:
        error:
          type: string
      type: object
...
    internal_httpapi.TwoFactorConfirmRequest:
      properties:
        code:
          type: string
        pendingToken:
          type: string
      type: object
    internal_httpapi.TwoFactorSetupRequest:
      properties:
        pendingToken:
          type: string
      type: object
...
    internal_httpapi.VerifyTwoFactorRequest:
      properties:
        code:
          type: string
        pendingToken:
          type: string
      type: object
...
    sudoStream_internal_auth.TwoFactorConfirmResult:
      properties:
        backupCodes:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
    sudoStream_internal_auth.TwoFactorSetup:
      properties:
        otpauthURL:
          type: string
        secret:
          type: string
      type: object
...
info:
  description: HTTP API for browsing and streaming media files.
  title: sudoStream API
  version: 0.35.0
openapi: 3.1.0
paths:
...
  /api/auth/2fa:
    delete:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_httpapi.DisableTwoFactorRequest'
                description: Password and TOTP code
                summary: body
        description: Password and TOTP code
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Unauthorized
      summary: Disable 2FA
      tags:
      - auth
  /api/auth/2fa/confirm:
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_httpapi.TwoFactorConfirmRequest'
                description: TOTP code
                summary: body
        description: TOTP code
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sudoStream_internal_auth.TwoFactorConfirmResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Unauthorized
      summary: Confirm 2FA setup
      tags:
      - auth
  /api/auth/2fa/setup:
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_httpapi.TwoFactorSetupRequest'
                description: Optional pending token
                summary: body
        description: Optional pending token
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sudoStream_internal_auth.TwoFactorSetup'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Unauthorized
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Conflict
      summary: Start 2FA setup
      tags:
      - auth
  /api/auth/2fa/verify:
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_httpapi.VerifyTwoFactorRequest'
                description: Pending token and code
                summary: body
        description: Pending token and code
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.AuthTokensResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Unauthorized
      summary: Verify 2FA login
      tags:
      - auth
  /api/auth/accept-invite:
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_httpapi.AcceptInviteRequest'
                description: Invite token
                summary: body
        description: Invite token
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Conflict
      summary: Accept invite
      tags:
      - auth
  /api/auth/change-email:
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_httpapi.ChangeEmailRequest'
                description: Email change
                summary: body
        description: Email change
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Unauthorized
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Internal Server Error
      summary: Request email change
      tags:
      - auth
  /api/auth/change-password:
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_httpapi.ChangePasswordRequest'
                description: Password change
                summary: body
        description: Password change
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.LoginResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Internal Server Error
      summary: Change password
      tags:
      - auth
  /api/auth/confirm-email:
    get:
      parameters:
      - description: Confirmation token
        in: query
        name: token
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.StatusResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
      summary: Confirm email
      tags:
      - auth
  /api/auth/forgot-password:
    post:
      description: Always returns 204 to avoid email enumeration.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_httpapi.ForgotPasswordRequest'
                description: Email
                summary: body
        description: Email
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
      summary: Forgot password
      tags:
      - auth
  /api/auth/login:
    post:
      description: Authenticates a user. May return a 2FA challenge instead of tokens.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_httpapi.LoginRequest'
                description: Credentials
                summary: body
        description: Credentials
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.LoginResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Internal Server Error
      summary: Login
      tags:
      - auth
  /api/auth/logout:
    post:
      description: Revokes the current refresh session.
      responses:
        "204":
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Unauthorized
      summary: Logout
      tags:
      - auth
  /api/auth/me:
    get:
      description: Returns the authenticated user profile.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.MeResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Unauthorized
      summary: Current user
      tags:
      - auth
  /api/auth/refresh:
    post:
      description: Reads the httpOnly refresh cookie and returns a new access token.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.RefreshResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Unauthorized
      summary: Refresh access token
      tags:
      - auth
...
servers:
- description: Local development server
  url: http://localhost:8080